f9twf.MdTreeOp.Get() 如果 Id=期貨複式單Id,但商品不存在,可能尚未收到行情資料,此時應建立此複式商品,這樣才能正常訂閱.

+++ fon9/ToStr.hpp
@@ -111,7 +111,7 @@ inline auto HEXToStrRev(char* pout, IntT value) -> decltype(HEXToStrRev(pout, st
 }
 
 inline char* PtrToStrRev(char* pout, const void* value) {
-   return HexToStrRev(pout, reinterpret_cast<uintmax_t>(value));
+   return HexToStrRev(pout, reinterpret_cast<uintptr_t>(value));
 }
 
 //--------------------------------------------------------------------------//
